requirements.txt
streamlit
python-docx
pdfplumber
xhtml2pdf